+++ HTTPLayerTests/HTTPLayerTests.swift
@@ -12,8 +12,8 @@ import XCTest
 class HTTPLayerTests: XCTestCase {
 
     override func setUp() {
-        HTTP.config.add(host: "https://jsonplaceholder.typicode.com", withContext: "", onKey: "DEFAULT")
-        HTTP.config.add(header: (key: "Content-type", value: "application/json; charset=UTF-8"), onKey: "DEFAULT")
+        HTTP.Config.add(host: "https://jsonplaceholder.typicode.com", withContext: "", onKey: "DEFAULT")
+        HTTP.Config.add(header: (key: "Content-type", value: "application/json; charset=UTF-8"), onKey: "DEFAULT")
     }
 
     override func tearDown() {
@@ -22,7 +22,7 @@ class HTTPLayerTests: XCTestCase {
 
     func testGetWithoutParameters() {
         let ex = expectation(description: "Expected server response, but naught was found")
-        HTTP.request.get(from: "/posts", withHostAndContext: "DEFAULT", andHeaders: "DEFAULT", receivingObjectType: [GETWithoutParametersResponse].self) { (response) in
+        HTTP.Request.get(from: "/posts", withHostAndContext: "DEFAULT", andHeaders: "DEFAULT", receivingObjectType: [GETWithoutParametersResponse].self) { (response) in
             switch response {
             case .success(let success):
                 XCTAssertNotNil(success)
@@ -40,7 +40,7 @@ class HTTPLayerTests: XCTestCase {
     
     func testGetWithQueryParameters() {
         let ex = expectation(description: "Expected server response, but naught was found")
-        HTTP.request.get(from: "/comments", usingQueryParameters: [ "postId" : "1" ], fromHostAndContext: "DEFAULT", andHeaders: "DEFAULT", receivingObjectType: [GETWithQueryParametersResponse].self) { (response) in
+        HTTP.Request.get(from: "/comments", usingQueryParameters: [ "postId" : "1" ], fromHostAndContext: "DEFAULT", andHeaders: "DEFAULT", receivingObjectType: [GETWithQueryParametersResponse].self) { (response) in
             switch response {
             case .success(let success):
                 XCTAssertNotNil(success)
@@ -58,7 +58,7 @@ class HTTPLayerTests: XCTestCase {
     
     func testGetWithPathParameters() {
         let ex = expectation(description: "Expected server response, but naught was found")
-        HTTP.request.get(from: "/posts", usingPathParameters: ["1", "comments"], fromHostAndContext: "DEFAULT", andHeaders: "DEFAULT", receivingObjectType: [GETWithPathParametersResponse].self) { (response) in
+        HTTP.Request.get(from: "/posts", usingPathParameters: ["1", "comments"], fromHostAndContext: "DEFAULT", andHeaders: "DEFAULT", receivingObjectType: [GETWithPathParametersResponse].self) { (response) in
             switch response {
             case .success(let success):
                 XCTAssertNotNil(success)
@@ -77,7 +77,7 @@ class HTTPLayerTests: XCTestCase {
     func testPostWithBodyParameter() {
         let ex = expectation(description: "Expected server response, but naught was found")
         let body = POSTWithBodyParameterRequest(title: "foo", body: "bar", userId: 1)
-        HTTP.request.post(to: "/posts", withBody: body, fromHostAndContext: "DEFAULT", andHeaders: "DEFAULT", receivingObjectType: POSTWithBodyParameterResponse.self) { (response) in
+        HTTP.Request.post(to: "/posts", withBody: body, fromHostAndContext: "DEFAULT", andHeaders: "DEFAULT", receivingObjectType: POSTWithBodyParameterResponse.self) { (response) in
             switch response {
             case .success(let success):
                 XCTAssertNotNil(success)
@@ -93,33 +93,10 @@ class HTTPLayerTests: XCTestCase {
         }
     }
     
-    //===================: Test Implemented but without a Service to Consume :==========================
-    //
-    //    func testPutWithBodyAndQueryParameters() {
-    //        let ex = expectation(description: "Expected server response, but naught was found")
-    //        let body = PUTWithBodyAndQueryParametersRequest(id: 1, title: "foo", body: "bar", userId: 1)
-    //        HTTP.request.put(on: "/posts", withBody: body, andQueryParameters: ["userId": "1"], fromHostAndContext: "DEFAULT", andHeaders: "DEFAULT", receivingObjectType: PUTWithBodyAndQueryParametersResponse.self) { (response) in
-    //            switch response {
-    //            case .success(let success):
-    //                XCTAssertNotNil(success)
-    //            case .failure(let failure):
-    //                XCTFail("Error: \(failure.localizedDescription)")
-    //            }
-    //            ex.fulfill()
-    //        }
-    //
-    //        waitForExpectations(timeout: 5) { (error) in
-    //            guard let error = error else { return }
-    //            XCTFail("Error: \(error.localizedDescription)")
-    //        }
-    //    }
-    //
-    //==================================================================================================
-    
     func testPutWithBodyAndPathParameters() {
         let ex = expectation(description: "Expected server response, but naught was found")
         let body = PUTWithBodyAndQueryParametersRequest(id: 1, title: "foo", body: "bar", userId: 1)
-        HTTP.request.put(on: "/posts", withBody: body, andPathParameters: ["1"], fromHostAndContext: "DEFAULT", andHeaders: "DEFAULT", receivingObjectType: PUTWithBodyAndQueryParametersResponse.self) { (response) in
+        HTTP.Request.put(on: "/posts", withBody: body, andPathParameters: ["1"], fromHostAndContext: "DEFAULT", andHeaders: "DEFAULT", receivingObjectType: PUTWithBodyAndQueryParametersResponse.self) { (response) in
             switch response {
             case .success(let success):
                 XCTAssertNotNil(success)
@@ -137,7 +114,7 @@ class HTTPLayerTests: XCTestCase {
     
     func testDeleteWithBodyAndPathParameters() {
         let ex = expectation(description: "Expected server response, but naught was found")
-        HTTP.request.delete(from: "posts", withBody: nil, andPathParameters: ["1"], fromHostAndContext: "DEFAULT", andHeaders: "DEFAULT", receivingObjectType: DELETEWithBodyAndPathParametersResponse.self) { (response) in
+        HTTP.Request.delete(from: "/posts", withPathParameters: ["1"], fromHostAndContext: "DEFAULT", andHeaders: "DEFAULT", receivingObjectType: DELETEWithPathParametersResponse.self) { (response) in
             switch response {
             case .success(let success):
                 XCTAssertNotNil(success)